requirements.txt
django
djangorestframework
pygments
httpie
pylint_django
factory-boy